+++ src/frontend/components/SøknadsSteg/Oppsummering/OppsummeringSteg/OmBarnaOppsummering.tsx
