+++ src/frontend/components/SøknadsSteg/Oppsummering/OppsummeringSteg/VelgBarnOppsummering.tsx
